Replace namshi/jose by lcobucci/jwt
