+++ Encoder/DefaultEncoder.php
@@ -2,71 +2,26 @@
 
 namespace Lexik\Bundle\JWTAuthenticationBundle\Encoder;
 
-use Lexik\Bundle\JWTAuthenticationBundle\Exception\JWTDecodeFailureException;
-use Lexik\Bundle\JWTAuthenticationBundle\Exception\JWTEncodeFailureException;
 use Lexik\Bundle\JWTAuthenticationBundle\Services\JWSProvider\JWSProviderInterface;
+use Namshi\JOSE\JWS;
 
 /**
  * Default Json Web Token encoder/decoder.
  *
  * @author Robin Chalas <robin.chalas@gmail.com>
+ *
+ * @deprecated since version 2.5, to be removed in 3.0
  */
-class DefaultEncoder implements HeaderAwareJWTEncoderInterface
+class DefaultEncoder extends LcobucciJWTEncoder
 {
-    /**
-     * @var JWSProviderInterface
-     */
-    protected $jwsProvider;
-
-    /**
-     * @param JWSProviderInterface $jwsProvider
-     */
     public function __construct(JWSProviderInterface $jwsProvider)
     {
-        $this->jwsProvider = $jwsProvider;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function encode(array $payload, array $header = [])
-    {
-        try {
-            $jws = $this->jwsProvider->create($payload, $header);
-        } catch (\InvalidArgumentException $e) {
-            throw new JWTEncodeFailureException(JWTEncodeFailureException::INVALID_CONFIG, 'An error occurred while trying to encode the JWT token. Please verify your configuration (private key/passphrase)', $e);
+        if (!class_exists(JWS::class)) {
+            throw new \RuntimeException('The "namshi/jose" library is deprecated, this bundle does not require it anymore. If you need to keep using it, require it in your composer.json.');
         }
 
-        if (!$jws->isSigned()) {
-            throw new JWTEncodeFailureException(JWTEncodeFailureException::UNSIGNED_TOKEN, 'Unable to create a signed JWT from the given configuration.');
-        }
+        @trigger_error(sprintf('The "%s\DefaultEncoder" class is deprecated since version 2.5 and will be removed in 3.0. Use "%s" or create your own encoder instead.', __NAMESPACE__, LcobucciJWTEncoder::class), E_USER_DEPRECATED);
 
-        return $jws->getToken();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function decode($token)
-    {
-        try {
-            $jws = $this->jwsProvider->load($token);
-        } catch (\Exception $e) {
-            throw new JWTDecodeFailureException(JWTDecodeFailureException::INVALID_TOKEN, 'Invalid JWT Token', $e);
-        }
-
-        if ($jws->isInvalid()) {
-            throw new JWTDecodeFailureException(JWTDecodeFailureException::INVALID_TOKEN, 'Invalid JWT Token');
-        }
-
-        if ($jws->isExpired()) {
-            throw new JWTDecodeFailureException(JWTDecodeFailureException::EXPIRED_TOKEN, 'Expired JWT Token');
-        }
-
-        if (!$jws->isVerified()) {
-            throw new JWTDecodeFailureException(JWTDecodeFailureException::UNVERIFIED_TOKEN, 'Unable to verify the given JWT through the given configuration. If the "lexik_jwt_authentication.encoder" encryption options have been changed since your last authentication, please renew the token. If the problem persists, verify that the configured keys/passphrase are valid.');
-        }
-
-        return $jws->getPayload();
+        parent::__construct($jwsProvider);
     }
 }
