+++ Resources/config/lcobucci.xml
@@ -5,12 +5,11 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="Lexik\Bundle\JWTAuthenticationBundle\Encoder\JWTEncoderInterface" alias="lexik_jwt_authentication.encoder.lcobucci" />
-        <service id="lexik_jwt_authentication.encoder.lcobucci" parent="lexik_jwt_authentication.encoder.abstract">
+        <service id="lexik_jwt_authentication.encoder.lcobucci" class="Lexik\Bundle\JWTAuthenticationBundle\Encoder\LcobucciJWTEncoder">
             <argument type="service" id="lexik_jwt_authentication.jws_provider.lcobucci" />
         </service>
 
-        <service id="Lexik\Bundle\JWTAuthenticationBundle\Services\JWSProvider\JWSProviderInterface" alias="lexik_jwt_authentication.jws_provider.default" />
+        <service id="Lexik\Bundle\JWTAuthenticationBundle\Services\JWSProvider\JWSProviderInterface" alias="lexik_jwt_authentication.jws_provider.lcobucci" />
         <service id="lexik_jwt_authentication.jws_provider.lcobucci" class="Lexik\Bundle\JWTAuthenticationBundle\Services\JWSProvider\LcobucciJWSProvider" public="false">
             <argument type="service" id="lexik_jwt_authentication.key_loader.raw"/>
             <argument>%lexik_jwt_authentication.encoder.crypto_engine%</argument>
