+++ Tests/Functional/app/config/autowiring.yml
@@ -10,5 +10,3 @@ lexik_jwt_authentication:
     secret_key: '%kernel.root_dir%/../config/jwt/private.pem'
     public_key: '%kernel.root_dir%/../config/jwt/public.pem'
     pass_phrase: testing
-    encoder:
-        service: lexik_jwt_authentication.encoder.default
